[KNight] Moved framework specific classes/interfaces to other package
KNight = Kotlin Night
