+++ common/src/main/kotlin/io/github/dector/knight/common/RandomUtils.kt
@@ -1,4 +1,4 @@
-package io.github.dector.quotes.common
+package io.github.dector.knight.common
 
 import java.util.*
 import java.util.concurrent.TimeUnit
@@ -13,7 +13,3 @@ fun <T> Array<T>.random(): T? = if (this.size > 0) this[Utils.rnd.nextInt(size)]
 fun <T> List<T>.random(): T? = if (this.size > 0) this[Utils.rnd.nextInt(size)] else null
 
 fun Int.randomUntil() =  if (this > 0) Utils.rnd.nextInt(this) else 0
-
-fun Date.add(time: Long, unit: TimeUnit) = Date(this.time + unit.toMillis(time))
-
-inline fun Int.minutesAsMillis(): Int = this * 60 * 1000
