+++ storage/src/main/kotlin/io/github/dector/quotes/repositories/CachedQuotesRepository.kt
@@ -1,5 +1,7 @@
 package io.github.dector.quotes.repositories
 
+import io.github.dector.knight.repositories.CachedRepository
+import io.github.dector.knight.repositories.ICacheStrategy
 import io.github.dector.quotes.domain.Quote
 import io.github.dector.quotes.storage.IStorableQuotesRepository
