+++ storage/src/main/kotlin/io/github/dector/quotes/repositories/IQuotesRepository.kt
@@ -1,5 +1,6 @@
 package io.github.dector.quotes.repositories
 
+import io.github.dector.knight.repositories.IRepository
 import io.github.dector.quotes.domain.Quote
 
 interface IQuotesRepository : IRepository<Quote>
