+++ usecases/src/main/kotlin/io/github/dector/quotes/usecases/IQuotesUseCase.kt
@@ -1,5 +1,6 @@
 package io.github.dector.quotes.usecases
 
+import io.github.dector.knight.usecases.IUseCase
 import io.github.dector.quotes.domain.Quote
 
 interface IQuoteUseCase : IUseCase<Quote, Throwable>
